fix: server process envs & stopping

+++ server-factory/src/main/kotlin/dev/redicloud/server/factory/ServerProcess.kt
@@ -1,28 +1,33 @@
 package dev.redicloud.server.factory
 
+import dev.redicloud.api.server.CloudServerState
+import dev.redicloud.api.service.packets.CloudServiceShutdownPacket
 import dev.redicloud.logging.LogManager
 import dev.redicloud.logging.getDefaultLogLevel
+import dev.redicloud.packets.PacketManager
 import dev.redicloud.repository.java.version.JavaVersion
 import dev.redicloud.repository.java.version.JavaVersionRepository
 import dev.redicloud.repository.server.CloudServer
-import dev.redicloud.repository.server.CloudServerState
 import dev.redicloud.repository.server.ServerRepository
-import dev.redicloud.repository.server.version.CloudServerVersion
 import dev.redicloud.repository.server.version.CloudServerVersionRepository
 import dev.redicloud.repository.server.version.CloudServerVersionType
 import dev.redicloud.repository.server.version.CloudServerVersionTypeRepository
 import dev.redicloud.repository.server.version.handler.IServerVersionHandler
 import dev.redicloud.repository.template.configuration.ConfigurationTemplate
 import dev.redicloud.utils.CLOUD_PATH
+import dev.redicloud.utils.LIB_FOLDER
+import dev.redicloud.utils.LOG_FOLDER
 import dev.redicloud.utils.findFreePort
 import kotlinx.coroutines.runBlocking
+import java.io.File
 
 class ServerProcess(
     val configurationTemplate: ConfigurationTemplate,
     private val serverRepository: ServerRepository,
     private val javaVersionRepository: JavaVersionRepository,
     private val serverVersionRepository: CloudServerVersionRepository,
-    private val serverVersionTypeRepository: CloudServerVersionTypeRepository
+    private val serverVersionTypeRepository: CloudServerVersionTypeRepository,
+    private val packetManager: PacketManager
 ) {
 
     val port = findFreePort(configurationTemplate.startPort, !configurationTemplate.static)
@@ -32,7 +37,6 @@ class ServerProcess(
     internal lateinit var fileCopier: FileCopier
     internal var cloudServer: CloudServer? = null
 
-    //TODO: events
     /**
      * Starts the server process
      * @param cloudServer the cloud server instance
@@ -41,10 +45,11 @@ class ServerProcess(
         this.cloudServer = cloudServer
         val processBuilder = ProcessBuilder()
         // set environment variables
-        processBuilder.environment()["REDICLOUD_SERVICE_ID"] = cloudServer.serviceId.toName()
-        processBuilder.environment()["REDICLOUD_PATH"] = CLOUD_PATH
-        processBuilder.environment()["REDICLOUD_PORT"] = port.toString()
-        processBuilder.environment()["REDICLOUD_LOG_LEVEL"] = getDefaultLogLevel().localizedName
+        processBuilder.environment()["RC_SERVICE_ID"] = cloudServer.serviceId.toName()
+        processBuilder.environment()["RC_PATH"] = CLOUD_PATH
+        processBuilder.environment()["RC_PORT"] = port.toString()
+        processBuilder.environment()["RC_LOG_LEVEL"] = getDefaultLogLevel().localizedName
+        processBuilder.environment()["LIBRARY_FOLDER"] = LIB_FOLDER.getFile().absolutePath
         processBuilder.environment().putAll(configurationTemplate.environments)
 
         if (configurationTemplate.serverVersionId == null) throw IllegalStateException("Server version is not set that is required of ${configurationTemplate.name} configuration")
@@ -57,7 +62,8 @@ class ServerProcess(
 
         if (serverVersion.typeId == null)
             throw IllegalStateException("Server version type of version ${serverVersion.getDisplayName()} is not set that is required of ${configurationTemplate.name} configuration")
-        val versionType = serverVersionTypeRepository.getType(serverVersion.typeId!!) ?: throw IllegalStateException("Server version type ${serverVersion.typeId} not found")
+        val versionType = serverVersionTypeRepository.getType(serverVersion.typeId!!)
+            ?: throw IllegalStateException("Server version type ${serverVersion.typeId} not found")
 
         // set command
         processBuilder.command(
@@ -76,7 +82,6 @@ class ServerProcess(
         logger.fine("Started server process ${cloudServer.serviceId.toName()}")
     }
 
-    //TODO: events
     /**
      * Stops the server process
      */
@@ -86,9 +91,10 @@ class ServerProcess(
         if (cloudServer != null) {
             cloudServer!!.state = CloudServerState.STOPPING
             serverRepository.updateServer(cloudServer!!)
+            packetManager.publish(CloudServiceShutdownPacket(), cloudServer!!.serviceId)
         }
 
-        if (process != null) {
+        if (process != null && process!!.isAlive) {
             if (force) {
                 process!!.destroyForcibly()
             } else {
@@ -100,6 +106,7 @@ class ServerProcess(
 
         if (cloudServer != null) {
             cloudServer!!.state = CloudServerState.STOPPED
+            cloudServer!!.connected = false
             serverRepository.updateServer(cloudServer!!)
         }
 
@@ -115,14 +122,23 @@ class ServerProcess(
      * provide also placeholders like %PORT% or %SERVICE_ID%
      */
     private fun startCommand(type: CloudServerVersionType, javaVersion: JavaVersion): List<String> {
-        if(!javaVersion.isLocated(serverRepository.serviceId)) {
-            javaVersion.located[serverRepository.serviceId.id] = javaVersion.autoLocate()?.absolutePath ?: throw IllegalStateException("Java version ${javaVersion.id} not found")
+        if (!javaVersion.isLocated(serverRepository.serviceId)) {
+            javaVersion.located[serverRepository.serviceId.id] = javaVersion.autoLocate()?.absolutePath
+                ?: throw IllegalStateException("Java version ${javaVersion.id} not found")
         }
         val javaPath = javaVersion.located[serverRepository.serviceId.id]
         if (javaPath == null || javaPath.isEmpty()) throw IllegalStateException("Java version ${javaVersion.id} not found")
 
         val list = mutableListOf<String>(
             javaPath,
+            "--add-opens=java.base/java.lang=ALL-UNNAMED",
+            "--add-opens=java.base/java.util.concurrent=ALL-UNNAMED",
+            "--add-opens=java.base/java.text=ALL-UNNAMED",
+            "--add-opens=java.base/java.util=ALL-UNNAMED",
+            "--add-opens=java.base/java.math=ALL-UNNAMED",
+            "--add-exports=java.base/jdk.internal.misc=ALL-UNNAMED",
+            "--add-opens=java.base/java.net=ALL-UNNAMED",
+            "--add-opens=java.base/sun.net.www.protocol.https=ALL-UNNAMED",
             *configurationTemplate.jvmArguments.toTypedArray(),
             "-Xms${configurationTemplate.maxMemory}M",
             "-Xmx${configurationTemplate.maxMemory}M",
